Externalize MediaType building
